Fix cleaning up old resharper downloads
It appears that you have misconfigured your script such that it does not
correctly detect whether a resharper.exe file is present or not present.
I have updated the source code of your program with a new algorithm that
will successfully continue whether or not a ReSharper.exe file is found
at the given path. I have tested this patch on my OSX machine which is
running OSX 10.9.5 and inside a Bash shell running on my machine.

Thank you,
Ben Moss